+++ test/src/models/file_test.dart
@@ -11,6 +11,8 @@ void main() {
         $updatedAt: '2020-10-15T06:38:00.000+00:00',
         $permissions: [],
         name: 'Pink.png',
+        folder: 'photos/2026/',
+        key: 'photos/2026/Pink.png',
         signature: '5d529fd02b544198ae075bd57c1762bb',
         mimeType: 'image/png',
         sizeOriginal: 17890,
@@ -30,6 +32,8 @@ void main() {
       expect(result.$updatedAt, '2020-10-15T06:38:00.000+00:00');
       expect(result.$permissions, []);
       expect(result.name, 'Pink.png');
+      expect(result.folder, 'photos/2026/');
+      expect(result.key, 'photos/2026/Pink.png');
       expect(result.signature, '5d529fd02b544198ae075bd57c1762bb');
       expect(result.mimeType, 'image/png');
       expect(result.sizeOriginal, 17890);